feat: 결제 취소 후 대기 상태의 결제를 결제하여 예약할 수 있도록 변경 및 패키지 리팩터링

+++ src/main/java/roomescape/dto/reservation/UserReservationRequest.java
@@ -1,6 +1,7 @@
 package roomescape.dto.reservation;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Objects;
 import roomescape.dto.payment.PaymentRequest;
@@ -11,7 +12,7 @@ public record UserReservationRequest(
         Long themeId,
         String paymentKey,
         String orderId,
-        int amount,
+        BigDecimal amount,
         String paymentType
 ) {
